+++ lib/exercise/domain/entities/writing_score_entity.dart
@@ -1,6 +1,6 @@
 class WritingScoreEntity {
   final bool isCorrect;
-  final int scorePercentage;
+  final double scorePercentage;
   final String feedback;
   final String performanceLevel;
